% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFilesInOne.R
\name{allFilesInOne}
\alias{allFilesInOne}
\title{Function to aggregate all the files (obs_raw, demo_raw, clincal_raw) in one file
and add additional information of interest, as the variant_misc assesed based on
the date and the country}
\usage{
allFilesInOne(
  obs_df,
  demo_df,
  clinical_df,
  variants_df,
  dateFormat,
  verbose = FALSE
)
}
\arguments{
\item{variants_df}{Data frame with the variant dates filtered by country in run_analysis()}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an on-time log from the function.}

\item{obs_raw}{Data frame with all the observations for the MISC patients}

\item{demo_raw}{Data frame with the MISC demographic information.}

\item{clincal_raw}{Data frame with all the clinical course details for MISC patients, starting with day 0 as the MISC admission date.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Given the three files (obs_raw, demo_raw and clinical_raw) readed with the
\code{readInput} file function, merge all in one and create as output a
\code{data.frame} object with all the data integrated, and adding the variant as
well as the specific week, month and year that will allow us to stratify the
analysis as required.
}
\examples{

misc_complete <- allFilesInOne(
                    obs_raw,
                    demo_raw,
                    clinical_raw,
                    dateFormat,
                    verbose
             )
}
