% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_sequences_hospitalisation.R
\name{count_sequences_hospitalisation}
\alias{count_sequences_hospitalisation}
\title{Function to estimate the sequences of hospitalization in Phase 2.2 data}
\usage{
count_sequences_hospitalisation(df, verbose = FALSE)
}
\arguments{
\item{df}{The clinical_raw data frame generated when applying the readInputFiles function.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an on-time log from the function.}
}
\value{
An object of class \code{data.frame} with two additional columns, the  \code{len_hospitalisation}
and \code{n_hospitalisation}.
}
\description{
Given the \code{df} LocalPatientClinicalCourse data use the in_hospital column
that generates two new columns, \code{n_hospitalisation} which represents which hospitalization,
and the \code{len_hospitalisation} which represents the number of consecutive days
the patient spent at the hospital for that hospitalization.
}
\examples{

count_sequences_hospitalisation(
              df      = "./" )
}
