% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data_for_stats.R
\name{format_data_for_stats}
\alias{format_data_for_stats}
\title{Function to get format the data for the statistical analysis}
\usage{
format_data_for_stats(integrated_df, raceAvailable, verbose)
}
\arguments{
\item{integrated_df}{The data.frame generated by the \code{allFilesInOne} function.}

\item{raceAvailable}{Value will be \code{TRUE} for the sites collecting this variable.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an on-time log from the function.}
}
\value{
An object of class \code{data.frame} formatted to perform statistical analysis and to generate table 1.
}
\description{
Given the \code{data.frame} output generated by the \code{allFilesInOne} function
generates as output a \code{data.frame} where each row represent a patient and
each column will represent each of the variables of interest for this study.
}
\examples{

formatted_output <- format_data_for_stats(
              integrated_df = misc_complete,
              raceAvailable,
              verbose
             )
}
