% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_table1.R
\name{misc_table1}
\alias{misc_table1}
\title{Function to get table 1 summary for MISC patients during the MISC hospitalization
Given the \code{data.frame} output generated by the \code{format_data_for_stats} function
generates as output the table 1 for the manuscript}
\usage{
misc_table1(
  formated_df,
  obfuscation_threshold,
  currSiteId,
  dir.output,
  verbose
)
}
\arguments{
\item{formated_df}{The data.frame generated by the \code{format_data_for_stats} function.}

\item{obfuscation_threshold}{FALSE if no obfuscation is needed, the numeric value when there is obfuscation}

\item{currSiteId}{Site ID}

\item{dir.output}{The path to the output directory where the aggregate counts and plots will be saved.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an on-time log from the function.}
}
\value{
Table 1 for the manuscript.
}
\description{
Function to get table 1 summary for MISC patients during the MISC hospitalization
Given the \code{data.frame} output generated by the \code{format_data_for_stats} function
generates as output the table 1 for the manuscript
}
\examples{

t1 <- misc_table1(
              formated_df,
              currSiteID,
              dir.output,
              verbose
             )
}
