% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_table2.R
\name{misc_table2}
\alias{misc_table2}
\title{Function to get table 2 summary for laboratory values of patients with MIS-C
Given the \code{data.frame} output generated by the \code{allFilesInOne} function
generates as output the table 2 for the manuscript}
\usage{
misc_table2()
}
\arguments{
\item{complete_df}{The data.frame generated by the \code{allFilesInOne} function.}

\item{obfuscation_threshold}{FALSE if no obfuscation is needed, the numeric value when there is obfuscation}

\item{currSiteId}{Site ID}

\item{dir.output}{The path to the output directory where the aggregate counts and plots will be saved.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an on-time log from the function.}
}
\value{
Table 2 for the manuscript.
}
\description{
Function to get table 2 summary for laboratory values of patients with MIS-C
Given the \code{data.frame} output generated by the \code{allFilesInOne} function
generates as output the table 2 for the manuscript
}
\examples{

t2 <- misc_table2(
              formated_df,
              currSiteID,
              dir.output,
              verbose
             )
}
