% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_table3.R
\name{misc_table3}
\alias{misc_table3}
\title{Function to get table 3 summary for outcomes
Given the \code{data.frame} output generated by the \code{allFilesInOne} function
generates as output the table 3 for the manuscript}
\usage{
misc_table3(
  complete_df,
  obfuscation_threshold,
  currSiteId,
  dir.output,
  raceAvailable,
  verbose
)
}
\arguments{
\item{complete_df}{The data.frame generated by the \code{allFilesInOne} function.}

\item{obfuscation_threshold}{FALSE if no obfuscation is needed, the numeric value when there is obfuscation}

\item{currSiteId}{Site ID}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an on-time log from the function.}
}
\value{
Table 3 for the manuscript.
}
\description{
Function to get table 3 summary for outcomes
Given the \code{data.frame} output generated by the \code{allFilesInOne} function
generates as output the table 3 for the manuscript
}
\examples{

t3 <- misc_table3(
              complete_df,
              currSiteId,
              obfuscation_threshold,
              verbose
             )
}
