% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{qc_summary}
\alias{qc_summary}
\title{Function to create a summary table of the MISC Phase 2.2 data}
\usage{
qc_summary(complete_df, obfuscation_threshold, during_misc_hosp = TRUE)
}
\arguments{
\item{complete_df}{The combined dataframe generated from the allFilesInOne function}

\item{obfuscation_threshold}{FALSE if no obfuscation is needed, the numeric value when there is obfuscation}

\item{during_misc_hosp}{By default \code{TRUE}, meaning that the summary will only be generated for data during the first MISC hospitalization.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an on-time log from the function.}
}
\value{
An object of class \code{data.frame} containing the summaries of labs, meds, and procedures described in the SRC
}
\description{
Given the \code{df} LocalPatientClinicalCourse data use the in_hospital column
that generates two new columns, \code{n_hospitalisation} which represents which hospitalization,
and the \code{len_hospitalisation} which represents the number of consecutive days
the patient spent at the hospital for that hospitalization.
}
\examples{

qc_summary(
              df      = "./" )
}
