% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInputFiles.R
\name{readInputFiles}
\alias{readInputFiles}
\title{Function to read the Phase 2.2 input files with patient-level data}
\usage{
readInputFiles(path, separator = ",", skip = 0, verbose = FALSE, ...)
}
\arguments{
\item{path}{A path were the input files are located.}

\item{separator}{The separator between columns (by default ",").}

\item{skip}{In case the file does not start with the column names, add the number of lines that should be skipped at the beggining.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an on-time log from the function.}
}
\value{
An object of class \code{list} with the \code{data.frames}.
}
\description{
Given the \code{path} where the files are located it will generates
a \code{list} with the \code{data.frame} object of the different files.
The input files that will be loaded are: LocalPatientSummary, LocalPatientObservation and LocalPatientClinicalcourse
For sites recording the race, we will also use an additional file: LocalPatientRace
}
\examples{

dataSet <- readInputFiles(
              path      = "./",
              separator = ",",
              skip      = 1,
             )
}
