% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnalysis.R
\name{runAnalysis}
\alias{runAnalysis}
\title{Function to run the 4CE MISC pediatric project using as input the 2.2 4CE misc cohort dataset}
\usage{
runAnalysis(
  dir.input,
  dir.output,
  obfuscation,
  raceAvailable,
  dateFormat,
  data_update_date,
  country,
  cbPalette = c("#999999", "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2",
    "#D55E00", "#CC79A7"),
  verbose
)
}
\arguments{
\item{dir.input}{The path where the 2.2 MISC cohort files are located.}

\item{dir.output}{The path to the output directory where the aggregate counts and plots will be saved.}

\item{obfuscation}{Determine the obfuscation threshold (FALSE if no obfuscation, or the numeric value of the obfuscation threshold if any). By default \code{FALSE}}

\item{raceAvailable}{set as TRUE or FALSE depending on whether the variable is being collected at your site. By default \code{TRUE}}

\item{dateFormat}{Specify the format of the date at your site (e.g., for "03-AUG-20", the format would be "\%d-\%b-\%y", see documentation](https://www.stat.berkeley.edu/~s133/dates.html))}

\item{data_update_date}{date at which the data has been updated in the local data warehouse. Used to estimate patient age at time of visit, since patients age in the 4CE demographic file is expected the age at data update.}

\item{country}{Specify the country of origin or your data (e.g, US, France, London, Spain)}

\item{cbPalette}{Color palette to use for the plots (set up by default)}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an on-time log from the function.}
}
\value{
An object of class \code{list} with the \code{data.frames}.
}
\description{
Information about how to create the 2.2 data can be found here: https://github.com/covidclinical/PhaseX.2SqlDataExtraction
The additional script required to create the MISC cohort can be found here: https://github.com/covidclinical/PhaseX.2SqlDataExtraction/blob/main/Extras/4CE_PhaseX.2_CustomCohorts_MISCPatients_mssql.sql
}
\examples{

FourCePhase2.2MISC::runAnalysis()
}
