% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_sequences_hospitalisation.R
\name{count_sequences_hospitalisation}
\alias{count_sequences_hospitalisation}
\title{Function to estimate the sequences of hospitalization in Phase 2.2 data}
\usage{
count_sequences_hospitalisation(df, ...)
}
\arguments{
\item{df}{The clinical_raw data frame generated when applying the readInputFiles function.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an on-time log from the function.}
}
\value{
An object of class \code{data.frame} with two additional columns, the  \code{len_hospitalisation}
and \code{n_hospitalisation}.
}
\description{
Given the \code{df} LocalPatientClinicalCourse data use the in_hospital column
that can get the values 1 when the patient is hospitalized, and 0 when the patient
is not hospitalized and generates two new columns, the \code{n_hospitalisation} one
that represents the number of distinct hospitalizations, and the \code{len_hospitalisation}
with the actual number of consecutive days the patient spent at the hospital.
}
\examples{

count_sequences_hospitalisation(
              df      = "./" )
}
